separate logger/level outside init function
