+++ geodiff/CMakeLists.txt
@@ -117,6 +117,8 @@ SET(geodiff_src
   src/geodiffrebase.hpp
   src/geodiffexporter.cpp
   src/geodiffexporter.hpp
+  src/geodifflogger.cpp
+  src/geodifflogger.hpp
   ${sqlite_src}
   ${libgpkg_src}
 )
@@ -142,8 +144,6 @@ ENDIF (BUILD_STATIC_LIBS)
 TARGET_COMPILE_DEFINITIONS(${GEODIFF_NAME} PUBLIC -DSQLITE_ENABLE_RTREE -DSQLITE_ENABLE_SESSION -DSQLITE_ENABLE_PREUPDATE_HOOK -DSQLITE_ENABLE_COLUMN_METADATA=1 -DSQLITE_MAX_VARIABLE_NUMBER=250000 -DSQLITE_ENABLE_FTS3=1 -DSQLITE_ENABLE_FTS3_PARENTHESIS=1 -DSQLITE_ENABLE_JSON1=1)
 TARGET_INCLUDE_DIRECTORIES(${GEODIFF_NAME} PRIVATE ${sqlite3_dir})
 TARGET_INCLUDE_DIRECTORIES(${GEODIFF_NAME} PRIVATE ${libgpkg_dir}/gpkg)
-
-
 
 # tests
 IF (ENABLE_TESTS)
